chore: rename common → utils
