+++ back/scripts/datasets/communities_financial_accounts.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pandas as pd
 
 from back.scripts.datasets.dataset_aggregator import DatasetAggregator
-from back.scripts.loaders.common import LOADER_CLASSES
+from back.scripts.loaders.utils import LOADER_CLASSES
 from back.scripts.utils.dataframe_operation import normalize_date
 
 LOGGER = logging.getLogger(__name__)
