+++ back/scripts/loaders/__init__.py
@@ -1,8 +1,8 @@
 from .base_loader import *  # noqa
-from .common import LOADER_CLASSES  # noqa
 from .csv_loader import *  # noqa
 from .excel_loader import *  # noqa
 from .json_loader import *  # noqa
 from .parquet_loader import *  # noqa
+from .utils import LOADER_CLASSES  # noqa
 from .xml_loader import *  # noqa
 from .zip_loader import *  # noqa
