+++ back/scripts/loaders/excel_loader.py
@@ -6,7 +6,7 @@ from typing import List, Optional, Union
 import pandas as pd
 
 from back.scripts.loaders.base_loader import BaseLoader
-from back.scripts.loaders.common import register_loader
+from back.scripts.loaders.utils import register_loader
 
 
 @register_loader
