+++ back/scripts/loaders/json_loader.py
@@ -5,7 +5,7 @@ from io import BytesIO
 import pandas as pd
 
 from back.scripts.loaders.base_loader import BaseLoader
-from back.scripts.loaders.common import register_loader
+from back.scripts.loaders.utils import register_loader
 
 
 @register_loader
